requirements.txt
jinja2>=3.1.4
pydantic>=2.8.2
httpx>=0.27.0
click>=8.1.7
python-dotenv>=1.0.1
pyyaml>=6.0.2
jsonschema>=4.23.0
rich>=13.9.2
